requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
biopython==1.85
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.40.41
botocore==1.40.41
build==1.3.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
choreographer==1.1.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.3.0
cloudpathlib==0.22.0
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
cymem==2.0.11
db-dtypes==1.4.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
durationpy==0.10
executing==2.2.1
fastapi==0.118.0
fastjsonschema==2.21.2
filelock==3.19.1
flasgger==0.9.7.1
Flask==3.1.2
flask-sock==0.7.0
flatbuffers==25.9.23
fqdn==1.5.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaleido==1.1.0
kubernetes==34.1.0
langcodes==3.5.0
language_data==1.3.0
lark==1.3.0
logistro==1.1.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.6.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.0
openai==2.1.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdfminer==20191125
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
pluggy==1.6.0
posthog==6.7.6
preshed==3.0.10
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
psycopg==3.2.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
PubChemPy==1.0.5
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
thinc==8.3.6
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0+cpu
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.37.0
uvloop==0.21.0
vanna==0.7.9
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
zipp==3.23.0
